#  Replace transitive apps and pods set with an iterable. 