@jeschkies 2017-12-08
## Master
## This Patch
| Benchmark                                       | (numberOfGroups) |(numberOfSavedApps) | Mode | Cnt| Score| Error |  Units
|---|---|---|---|---|---|-----|---|
| assignDynamicServicePorts | 5 |  100 | avgt | 10 | 0.036| ±0.002|  ms/op
| assignDynamicServicePorts | 5 |  500 | avgt | 10 | 0.141| ±0.006|  ms/op
| assignDynamicServicePorts | 5 | 1000 | avgt | 10 | 0.271| ±0.018|  ms/op
| assignDynamicServicePorts | 5 | 2500 | avgt | 10 | 0.647| ±0.036|  ms/op
| assignDynamicServicePorts | 5 | 5000 | avgt | 10 | 1.515| ±0.406|  ms/op
| assignDynamicServicePorts | 5 |10000 | avgt | 10 | 3.585| ±0.148|  ms/op
| assignDynamicServicePorts |15 |  100 | avgt | 10 | 0.036| ±0.001|  ms/op
| assignDynamicServicePorts |15 |  500 | avgt | 10 | 0.135| ±0.003|  ms/op
| assignDynamicServicePorts |15 | 1000 | avgt | 10 | 0.262| ±0.006|  ms/op
| assignDynamicServicePorts |15 | 2500 | avgt | 10 | 0.591| ±0.046|  ms/op
| assignDynamicServicePorts |15 | 5000 | avgt | 10 | 1.591| ±0.222|  ms/op
| assignDynamicServicePorts |15 |10000 | avgt | 10 | 3.304| ±0.175|  ms/op