@jeschkies 2017-12-08
# Benchmark `buildRootGroup`
`sbt "benchmark/jmh:run -i 10 -wi 3 -f1 -t1 .*RootGroupBenchmark.buildRootGroup"`
## Master

## This Patch
|(appsPerGroup)| (groupDepth) | Mode | Cnt | Score | Error | Units
|---|---|---|---|---|-----|---|
|   2|  5| avgt | 10|  0.054| ± 0.005| ms/op
|   2| 10| avgt | 10|  0.172| ± 0.021| ms/op
|   2| 20| avgt | 10|  0.461| ± 0.013| ms/op
|  10|  5| avgt | 10|  0.264| ± 0.031| ms/op
|  10| 10| avgt | 10|  0.736| ± 0.085| ms/op
|  10| 20| avgt | 10|  1.949| ± 0.055| ms/op
| 100|  5| avgt | 10|  2.318| ± 0.214| ms/op
| 100| 10| avgt | 10|  6.948| ± 1.484| ms/op
| 100| 20| avgt | 10| 20.625| ± 2.231| ms/op
|1000|  5| avgt | 10| 27.452| ± 1.402| ms/op
|1000| 10| avgt | 10| 76.132| ± 4.700| ms/op
|1000| 20| avgt | 10|320.854| ±20.372| ms/op