@jeschkies 2017-12-08
@zen-dog could you check the dependent PRs #5804 and #5805 as well. Thanks 🙂.